+++ src/xrGame/WeaponMagazined.h
@@ -197,7 +197,7 @@ public:
 
 	virtual void SetFireMode(int mode)
 	{
-		if (mode >= m_aFireModes.size()) mode = 0;
+		if ((u32)mode >= m_aFireModes.size()) mode = 0;
 		m_iCurFireMode = mode;
 		SetQueueSize(GetCurrentFireMode());
 	};
